Decouple from Decoder (#5)
This decouples the main decoding logic from the Decoder interface. The
requirements of the Decoder interface are now encoded in the
`unmarshaler` struct.

This will make it possible to customize the Decoder interface in the
future.

To keep things clean, the Decoder type and friends were moved to their
own file.